QR/requirements.txt
segno==1.6.1
Pillow==11.0.0
Flask==3.0.3
